Добавила удаление элементов из разметки при отсутствии данных
