+++ js/main.js
@@ -1,4 +1,4 @@
 import {createSimilarRentObjects as data} from './data.js';
-import {createLayoutForSimularRenObjects} from './layout.js';
+import {createLayoutForRentObject} from './layout.js';
 
-createLayoutForSimularRenObjects(data, 1);
+createLayoutForRentObject(data[0]);
